\name{is.3dpoints}
\alias{is.3dpoints}
\title{Spatio-temporal point objects}

\description{
Tests for data in spatio-temporal point format.}

\usage{
is.3dpoints(x)
}

\arguments{
  \item{x}{any object.}
 }

\value{
is.3dpoints returns TRUE if x is a spatio-temporal points object, FALSE
otherwise. }

\author{Edith Gabriel <edith.gabriel@univ-avignon.fr>, Peter Diggle, Barry
Rowlingson.}
